+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 7.63
 -----
 - Show the Up Next bar on the tab bar. [#1613]
+- New design for the Podcasts grid layouts. [#1628]
 
 7.62
 -----