+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 -----
 - Show the Up Next bar on the tab bar. [#1613]
 - New design for the Podcasts grid layouts. [#1628]
+- Shake the device to restart Sleep Timer [#1627]
+- Playback fades out when finishing a sleep timer [#1629]
 
 7.62
 -----